requirements.txt
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
pyserial>=3.5
flask>=2.3.0
flask-cors>=4.0.0
